Added redirect hostname discovery plugin with optional /etc/hosts update and duplicate checks in /etc/hosts
